organised components (groupped them under folders), added comments, adjusted styling on responsive design
