+++ src/App.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import AdminPage from "./components/AdminPage";
-import CheckoutPage from "./components/CheckoutPage";
+import CheckoutPage from "./components/cart-checkout/CheckoutPage";
 import DetailPage from "./components/DetailPage";
 import Layout from "./components/Layout";
 import NotFound from "./components/NotFound";
-import OrderConfirmation from "./components/OrderConfirmation";
+import OrderConfirmation from "./components/cart-checkout/OrderConfirmation";
 import ProductPage from "./components/ProductPage";
 import EmptyPage from "./components/shared/EmptyPage";
 import StartPage from "./components/StartPage";
