+++ src/components/shared/HomeButton.tsx
@@ -17,7 +17,6 @@ const HomeButton = (props: Props) => {
         width: "fit-content",
         backgroundColor: "#CAC2B9",
         margin: "1rem auto",
-        textTransform: "none",
       }}
     >
       {props.message}
