+++ src/context/CartContextProvider.tsx
@@ -57,6 +57,64 @@ const CartProvider: FC = (props) => {
   const [isSwish, setIsSwish] = useState<Boolean>(false);
   const [isInvoice, setIsInvoice] = useState<Boolean>(false);
 
+  /** add items to cart: if the item does not exist in the cart, add; otherwise increase quantity by 1 */
+  const addToCart = async (product: ProductData) => {
+    if (cart.some((item) => item.id === product.id)) {
+      const updatedCart = cart.map((item) => {
+        if (product.id !== item.id) return item;
+        return { ...item, quantity: item.quantity + 1 };
+      });
+      setCart(updatedCart);
+    } else {
+      const cartItem: ItemData = { ...product, quantity: 1 };
+      setCart([...cart, cartItem]);
+    }
+  };
+
+  /** add item quantity by 1 in shopping cart */
+  const onAddQuantity = (product: ItemData) => {
+    const updatedQuantity = cart.map((item) => {
+      if (product.id !== item.id) return item;
+      return { ...item, quantity: item.quantity + 1 };
+    });
+    setCart(updatedQuantity);
+  };
+
+  /** reduce item quantity by 1 in shopping cart */
+  const onReduceQuantity = (product: ItemData) => {
+    const updatedQuantity = cart.map((item) => {
+      if (product.id === item.id && item.quantity > 1)
+        return { ...item, quantity: item.quantity - 1 };
+      return item;
+    });
+    setCart(updatedQuantity);
+  };
+
+  /** remove item from from the cart */
+  const removeFromCart = (product: ItemData) => {
+    if (cart.find((item) => item.id === product.id)) {
+      const updatedCart = cart.filter((item) => item.id !== product.id);
+      setCart(updatedCart);
+    }
+  };
+
+  /** set cart to empty array */
+  const emptyCart = () => {
+    setCart([]);
+  };
+
+  /** set state when a shipper is selected */
+  const selectShippment = (provider: ShippingProvider) => {
+    setShipper(provider);
+  };
+
+  /** set state when a payment method is selected */
+  const selectPaymentMethod = (method: String) => {
+    setPaymentMethod(method);
+  };
+
+  /** the below 3 function are for changing the state of the specific payment method when it is being selected,
+   *  solely for the form's conditional validation */
   const selectSwish = () => {
     setIsSwish(true);
     setIsCreditCard(false);
@@ -73,55 +131,6 @@ const CartProvider: FC = (props) => {
     setIsSwish(false);
     setIsCreditCard(false);
     setIsInvoice(true);
-  };
-
-  const addToCart = async (product: ProductData) => {
-    if (cart.some((item) => item.id === product.id)) {
-      const updatedCart = cart.map((item) => {
-        if (product.id !== item.id) return item;
-        return { ...item, quantity: item.quantity + 1 };
-      });
-      setCart(updatedCart); // update to LS
-    } else {
-      const cartItem: ItemData = { ...product, quantity: 1 };
-      setCart([...cart, cartItem]);
-    }
-  };
-
-  const onAddQuantity = (product: ItemData) => {
-    const updatedQuantity = cart.map((item) => {
-      if (product.id !== item.id) return item;
-      return { ...item, quantity: item.quantity + 1 };
-    });
-    setCart(updatedQuantity);
-  };
-
-  const onReduceQuantity = (product: ItemData) => {
-    const updatedQuantity = cart.map((item) => {
-      if (product.id === item.id && item.quantity > 1)
-        return { ...item, quantity: item.quantity - 1 };
-      return item;
-    });
-    setCart(updatedQuantity);
-  };
-
-  const removeFromCart = (product: ItemData) => {
-    if (cart.find((item) => item.id === product.id)) {
-      const updatedCart = cart.filter((item) => item.id !== product.id);
-      setCart(updatedCart);
-    }
-  };
-
-  const emptyCart = () => {
-    setCart([]);
-  };
-
-  const selectShippment = (provider: ShippingProvider) => {
-    setShipper(provider);
-  };
-
-  const selectPaymentMethod = (method: String) => {
-    setPaymentMethod(method);
   };
 
   return (
